backend/requirements.txt
fastapi
uvicorn
redis[hiredis]
pydantic[email]
pydantic-redis
psycopg2-binary
sqlalchemy
python-dotenv
phonenumbers
pyjwt
twilio
bcrypt
alembic